[remoteExec] Add first working prototype of remote executor
The rationale is that the node that runs the JS shell will run
the remote executor program, and the manager node manages the fuzzing
corpus and other fuzzing logic (eg. mutations) while communicating
with the executor node via network (see remote-executor-protocol.h).

This commit only includes a working remote executor implementation and
a test-executor.c binary to test that the remote executor works as
expected (although it is not a comprehensive test suite). Further
work is needed to modify the manager node to communicate to the
executor node via network.

Only REPRL commands are implemented in the remote executor protocol
layer. Coverage commands are missing in this commit.

It is currently only tested on CheriBSD.