[remoteExec] Integrate the remote execution protocol
This builds on the previous remote executor implementation to add command packets
for coverage instructions (such as coverage bitmap reset, clearing edge data...). Then,
it integrates this into the manager side so that the manager node (running on Linux,
for example) communicates to the remote executor (running on CheriBSD, for example)
via network. To do this, relevant REPRL and coverage functions are replaced with remote
versions of these functions if the --remoteExecutor argument is set.